consolidate setup code for relay_controller tests
