+++ lib/cog/support/model_utilities.ex
@@ -189,6 +189,51 @@ defmodule Cog.Support.ModelUtilities do
   end
 
   @doc """
+  Create a command with the given name
+  """
+  def command(name) do
+    bundle = case Repo.get_by(Bundle, name: "cog") do
+      nil ->
+        bundle("cog")
+      bundle ->
+        bundle
+    end
+
+    %Command{}
+    |> Command.changeset(%{name: name, bundle_id: bundle.id})
+    |> Repo.insert!
+  end
+
+  @doc """
+  Creates a bundle record
+  """
+  def bundle(name, commands \\ %{"echo": %{"executable" => "/bin/echo"}}, opts \\ []) do
+
+    bundle_template = %{
+      "name" => name,
+      "version" => "0.1.0",
+      "cog_bundle_version" => 2,
+      "commands" => commands
+    }
+
+    bundle_config = Enum.into(opts, bundle_template, fn
+      ({key, value}) when is_atom(key) ->
+        {Atom.to_string(key), value}
+      (opt) ->
+        opt
+    end)
+
+    bundle = %Bundle{}
+    |> Bundle.changeset(%{name: name, version: bundle_config["version"], config_file: bundle_config})
+    |> Repo.insert!
+
+    namespace(name, bundle.id)
+
+    bundle
+  end
+
+
+  @doc """
   Creates a relay record
   """
   def relay(name, token, opts \\ []) do
@@ -228,6 +273,22 @@ defmodule Cog.Support.ModelUtilities do
     |> RelayGroupAssignment.changeset(%{group_id: group_id,
                                         bundle_id: bundle_id})
     |> Repo.insert!
+  end
+
+  @doc """
+  Creates a relay, relay-group and bundle. Then assigns the bundle and adds the
+  relay to the relay-group
+  """
+  def create_relay_bundle_and_group(name, opts \\ []) do
+    relay = relay("relay-#{name}",
+                  Keyword.get(opts, :token, "sekrit"),
+                  Keyword.get(opts, :relay_opts, []))
+    bundle = bundle("bundle-#{name}")
+    relay_group = relay_group("group-#{name}")
+    add_relay_to_group(relay_group.id, relay.id)
+    assign_bundle_to_group(relay_group.id, bundle.id)
+
+    {relay, bundle, relay_group}
   end
 
   @doc """
